changed instruction counter to ulong so it can run for > 50k years

+++ Model/Interfaces/ICPU.cs
@@ -30,7 +30,7 @@ public interface ICPU
 
     void SetBreakpoint(Breakpoint breakpoint);
 
-    int InstructionCounter { get; }
+    ulong InstructionCounter { get; }
 
     void Attach(IRK8E peripheral);
